dosextend: new program
Add functionality to extend a DOS partition table
with an additional partition that encompasses the
remaining portion of a disk. (This is performed
without exec-ing sfdisk(8)).

Since we've got built-in support for reading DOS
partition tables now, don't use sfdisk(8) for
gptimage when '-d' is passed.

(GPT support is forthcoming.)

Also, improve testing.